This will probably introduce some weird bugs, but I'm hoping not.  If it does contact me and I promise I'll hunt em down and capture and or kill them, just like the terrorists...
The code no longer reloads all of the modules all the time, and instead calls
the _Load() functions on the payloads when it has to, and hopefully all of
the other modules will behave ok.  I think the nops/encoders are getting
loaded everytime anyway, since they get called with like MakeNop, etc, but
I'd have to double check that.  Anyway, hopefully this works...